use testify/suite for windows too
